Jenkins CFN.
Cloud Formation Template to create a jenkins server with EIP to access it.
Note: Parameters has to be changed according to environment. After jenkins server is launched use the url in stack outputs to access jenkins. One time process of manual login to jenkins server is needed to update intial credentials and restart jenkins service.